TCN/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_tflow_select=2.1.0=gpu
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.7.4.post0=py38h7f8727e_2
aiohttp-cors=0.7.0=py_0
aioredis=1.3.1=pyhd3eb1b0_0
aiosignal=1.2.0=pyhd3eb1b0_0
astor=0.8.1=py38h06a4308_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py38h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blessings=1.7=py38h578d9bd_1005
blinker=1.4=py38h06a4308_0
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py38h06a4308_0
cffi=1.14.6=py38h400218f_0
chardet=4.0.0=py38h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
colorama=0.4.4=pyh9f0ad1d_0
colorful=0.5.4=pyhd8ed1ab_0
cryptography=3.4.8=py38hd23ed53_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cupti=10.1.168=0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.0=pyhd3eb1b0_0
deprecated=1.2.13=pyh6c4a22f_0
entrypoints=0.3=py38_0
expat=2.4.1=h2531618_2
ffmpeg=4.2.2=h20bf706_0
filelock=3.4.0=pyhd8ed1ab_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py38h7f8727e_0
fsspec=2021.10.1=pyhd3eb1b0_0
future=0.18.2=py38_1
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
#google-api-core=2.3.2=pyhd8ed1ab_0
#google-auth=1.33.0=pyhd3eb1b0_0
#google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
#googleapis-common-protos=1.53.0=py38h578d9bd_1
gpustat=0.6.0=pyhd8ed1ab_1
grpcio=1.42.0=py38hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py38hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
hiredis=1.1.0=py38h27cfd23_1
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.8.2=py38h06a4308_0
importlib_resources=5.4.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
#ipykernel=6.6.0=py38he5a9106_0
#ipython=7.29.0=py38hb070fc8_0
jedi=0.18.0=py38h06a4308_1
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jsonschema=4.3.2=pyhd8ed1ab_0
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_core=4.9.1=py38h06a4308_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py38h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libllvm11=11.1.0=h3826bc1_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvmlite=0.37.0=py38h295c915_1
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py38h06a4308_0
matplotlib=3.5.0=py38h06a4308_0
matplotlib-base=3.5.0=py38h3ed280b_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
msgpack-python=1.0.2=py38hff7bd54_1
multidict=5.1.0=py38h27cfd23_2
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=py38hd09550d_3
#numba=0.54.1=py38h51133e4_0
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
nvidia-ml=7.352.0=py_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=pyhd3eb1b0_0
#opencensus=0.8.0=pyhd8ed1ab_1
opencensus-context=0.1.2=py38h578d9bd_5
openh264=2.1.1=h4ff587b_0
openssl=1.1.1l=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.4=py38h8c16a72_0
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
prometheus_client=0.12.0=pyhd8ed1ab_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
#protobuf=3.17.2=py38h295c915_0
psutil=5.8.0=py38h497a2fe_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=py38h497a2fe_0
pydeprecate=0.3.1=pyhd8ed1ab_0
#pygments=2.10.0=pyhd3eb1b0_0
pyjwt=2.1.0=py38h06a4308_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.8.0=py3.8_cuda10.1_cudnn7.6.3_0
#pytorch-lightning=1.5.7=pyhd8ed1ab_0
pytorch-model-summary=0.1.1=py_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=py38h7f8727e_1
pyzmq=22.3.0=py38h295c915_2
qt=5.9.7=h5867ecd_1
#ray-core=1.4.0=py38h295c915_1
#ray-tune=1.4.0=py38h06a4308_1
readline=8.1=h27cfd23_0
redis-py=4.0.2=pyhd8ed1ab_0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.0.1=py38h51133e4_0
scipy=1.7.1=py38h292c36d_2
setproctitle=1.1.10=py38h497a2fe_1004
setuptools=58.0.4=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tabulate=0.8.9=pyhd8ed1ab_0
tbb=2021.4.0=hd09550d_0
#tensorboard=2.4.0=pyhc547734_0
#tensorboard-plugin-wit=1.6.0=py_0
#tensorboardx=2.4=pyhd8ed1ab_0
#tensorflow=2.4.1=gpu_py38h8a7d6ce_0
#tensorflow-base=2.4.1=gpu_py38h29c2da4_0
#tensorflow-estimator=2.6.0=pyh7b7c402_0
# tensorflow-gpu=2.4.1=h30adc30_0
termcolor=1.1.0=py38h06a4308_1
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.8.0=py38
#torchmetrics=0.6.2=pyhd8ed1ab_0
# torchsummary=1.5.1=pypi_0      # 单独通过pip install安装
torchvision=0.9.0=py38_cu101
tornado=6.1=py38h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd3eb1b0_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
wrapt=1.13.3=py38h7f8727e_2
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py38h27cfd23_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
